hello-world/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.13.0
MarkupSafe==1.1.1
requests==2.22.0
urllib3==1.25.3
Werkzeug==0.15.5
